backend/requirements.txt
Flask
flask-cors
flask-mail
python-dotenv
mysql-connector-python
openai
gunicorn